Create a trait ShuffledJoin for SMJ and SHJ
